Create test_process_command.py test
